[wip] Make source not lazy for geojson source
Use common initialization logic to create an unique pointer to an mbgl
source object, up front, when a MGL source is created. Keep a raw
pointer to the unique pointer that is pointed at the mbgl source
instance when a MGL source is created or when a MGL source is
obtained by identifier from MGLStyle.

Once the transfer of ownership of the mbgl source takes place, the
unique ptr is null. The raw pointer can be used, internally, for
future work that involves mutating the source.